extract sign and poll methods from Wallet into Wallet+Transactions extension
